welcome circle changes and role permissions
