+++ src/handlers/index.js
@@ -2,12 +2,14 @@ const addAddressHandler = require('./addAddress')
 const helpHandler = require('./help')
 const tokenHandler = require('./token')
 const scoreHandler = require('./score')
+const introHandler = require('./intro')
 
 const handlers = new Map([
   ['addaddress', addAddressHandler],
   ['help', helpHandler],
   ['token', tokenHandler],
-  ['score', scoreHandler]
+  ['score', scoreHandler],
+  ['intro', introHandler],
 ])
 
 module.exports = handlers
